+++ src/sparql-notebook-controller.ts
@@ -6,7 +6,7 @@ export class SparqlNotebookController {
   readonly controllerId = "sparql-notebook-controller-id";
   readonly notebookType = "sparql-notebook";
   readonly label = "Sparql Notebook";
-  readonly supportedLanguages = ["sparql", "rdfmapping"];
+  readonly supportedLanguages = ["sparql"];
 
   private readonly _controller: vscode.NotebookController;
   private _executionOrder = 0;
@@ -37,16 +37,25 @@ export class SparqlNotebookController {
     const execution = this._controller.createNotebookCellExecution(cell);
     execution.executionOrder = ++this._executionOrder;
     execution.start(Date.now()); // Keep track of elapsed time to execute cell.
-    if (globalConnection.connection === null) {
-      vscode.window.showErrorMessage("Not connected to a SPARQL Endpoint");
-      execution.end(true, Date.now());
-      return;
+
+    let client: SparqlClient | null = null;
+    const sparqlQuery = cell.document.getText();
+    const documentEndpoint = this._getEndpointFromQuery(sparqlQuery);
+    if (documentEndpoint) {
+      client = new SparqlClient(documentEndpoint, "", "");
+    } else {
+      if (globalConnection.connection === null) {
+        vscode.window.showErrorMessage("Not connected to a SPARQL Endpoint");
+        execution.end(true, Date.now());
+        return;
+      }
+      client = new SparqlClient(
+        globalConnection.connection.data.endpointURL,
+        globalConnection.connection.data.user,
+        globalConnection.connection.data.passwordKey
+      );
     }
-    const client = new SparqlClient(
-      globalConnection.connection.data.endpointURL,
-      globalConnection.connection.data.user,
-      globalConnection.connection.data.passwordKey
-    );
+
     const queryResult = await client
       .query(cell.document.getText())
       .catch((error) => {
@@ -81,9 +90,11 @@ export class SparqlNotebookController {
 
   private _writeTurtleResult(resultTTL: string): vscode.NotebookCellOutput {
     return new vscode.NotebookCellOutput([
-      // vscode.NotebookCellOutputItem.text(resultTTL, "application/x-turtle"),
-      vscode.NotebookCellOutputItem.text(resultTTL, "text/x-turtle"),
       vscode.NotebookCellOutputItem.text(resultTTL, "text/plain"),
+      vscode.NotebookCellOutputItem.text(
+        `\`\`\`turtle\n${resultTTL}\n\`\`\``,
+        "text/markdown"
+      ),
     ]);
   }
 
@@ -173,6 +184,24 @@ tbody tr:hover {
         message: error.message,
       }),
     ]);
+  }
+
+  private _getEndpointFromQuery(sparqlQuery: string): string | undefined {
+    const commentLines = sparqlQuery
+      .split("\n")
+      .map((l) => l.trim())
+      .filter((l) => l.startsWith("#"));
+    const endpointExp = /\[endpoint=(.*)\]/gm;
+    const endpoints: string[] = [];
+    commentLines.every((comment: string) => {
+      const match = endpointExp.exec(comment);
+      if (match) {
+        endpoints.push(match[1]);
+        return false;
+      }
+      return true;
+    });
+    return endpoints.shift();
   }
 
   dispose() {}